backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
alembic>=1.14.0
asyncpg>=0.30.0
pydantic-settings>=2.6.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.9.0
email-validator>=2.1.0
